+++ src/SentimentScorerSentences.h
@@ -10,10 +10,12 @@ struct SentimentScorerSentences : public RcppParallel::Worker {
   const std::unordered_map< std::string, double > valenceMap;
   const std::string how;
   const int nL;
+  const int N;
   std::unordered_map< int, std::unordered_map< std::string, double> > frequencyMap;
   std::unordered_map< std::string, double > inverseFrequencyMap;
   const bool isFreqWeighting;
-  const bool hasValenceShifters;
+  const int valenceType;
+
   // output
   RcppParallel::RMatrix<double> sentScores;
 
@@ -22,12 +24,13 @@ struct SentimentScorerSentences : public RcppParallel::Worker {
                            const std::unordered_map< std::string, double > valenceMap,
                            const std::string how,
                            int nL,
+                           int N,
                            std::unordered_map< int, std::unordered_map< std::string, double > > frequencyMap,
                            std::unordered_map< std::string, double > inverseFrequencyMap,
                            const bool isFreqWeighting,
-                           const bool hasValenceShifters,
+                           const int valenceType,
                            Rcpp::NumericMatrix sentScores)
-    : texts(texts), lexiconMap(lexiconMap), valenceMap(valenceMap), how(how), nL(nL), frequencyMap(frequencyMap), inverseFrequencyMap(inverseFrequencyMap), isFreqWeighting(isFreqWeighting), hasValenceShifters(hasValenceShifters), sentScores(sentScores) {}
+    : texts(texts), lexiconMap(lexiconMap), valenceMap(valenceMap), how(how), nL(nL), N(N), frequencyMap(frequencyMap), inverseFrequencyMap(inverseFrequencyMap), isFreqWeighting(isFreqWeighting), valenceType(valenceType), sentScores(sentScores) {}
 
   void operator()(std::size_t begin, std::size_t end) {
 
@@ -50,41 +53,50 @@ struct SentimentScorerSentences : public RcppParallel::Worker {
 
       for (int j = 0; j < nTokens; j++) {
         std::string token = tokens[j];
-        check_for_commas(token, nPuncts, punctPosition, j, tokens);
-
         double tokenFrequency = 1.0, tokenInverseFrequency = 1.0;
-
+        if (isFreqWeighting) {
+          update_token_frequency(tokenFrequency, freqMap, token);
+          update_token_inverse_frequency(tokenInverseFrequency, inverseFrequencyMap, token, how);
+        }
         if (lexiconMap.find(token) != lexiconMap.end()) { // hit
           tokenScores[j] = lexiconMap.at(token);
-          std::vector<int> shifters(4);
 
-          if (how != "proportional" && how != "counts" && how != "squareRootCounts") {
-            update_token_weights(tokenWeights, normalizer, nPolarized, j, nTokens, how, nL, tokenScores, tokenFrequency, tokenInverseFrequency, maxTokenFrequency); //step 3 and 4: get token Weights
+          if (how != "proportional" && how != "counts" && how != "proportionalSquareRoot") {
+            update_token_weights(tokenWeights, normalizer, nPolarized, j, nTokens, how, nL, tokenScores, tokenFrequency, tokenInverseFrequency, maxTokenFrequency, N);
           }
 
-          if (hasValenceShifters) {
-            int st = 0, en = 0, position = -1;
+          if (valenceType == 1) { // bigrams approach
+            int k = std::max(0, j - 1);
+            if (valenceMap.find(tokens[k]) != valenceMap.end()) { // bigram valence shifting
+              tokenShifters[j] = valenceMap.at(tokens[k]);
+            }
+          } else if (valenceType == 2) { // clusters approach
+            check_for_commas(token, nPuncts, punctPosition, j, tokens);
+            int st = 0, en = 0;
+            std::vector<int> shifters(4);
             set_cluster_bounds(st, en, nPuncts, punctPosition, nTokens, tokens, lB, nB, nA, j);
-
             for (int k = st; k < en; k++) {
               if (k == j) continue;
-
               std::string token_k = tokens[k];
-              if (valenceMap.find(token_k) != valenceMap.end()) {
+              if (lexiconMap.find(token_k) != lexiconMap.end()) {
+                tokenScores[k] = lexiconMap.at(token_k);
+                if (how != "proportional" && how != "counts" && how != "proportionalSquareRoot") {
+                  update_token_weights(tokenWeights, normalizer, nPolarized, k, nTokens, how, nL, tokenScores, tokenFrequency, tokenInverseFrequency, maxTokenFrequency, N);
+                }
+              } else if (valenceMap.find(token_k) != valenceMap.end()) {
                 double valType = valenceMap.at(token_k);
+                int position = 0; // initialize
                 if (valType == 4) {
-                  position = 0;
-                  if (j < k) position = 1;
+                  if (k > j) position = 1; // adv. conjunction after hit (position = 0 if before)
                 }
                 update_primary_shifters_sentence(shifters, valType, position);
               }
             }
-            tokenShifters[j] = compute_sentence_impact(shifters, position);
+            tokenShifters[j] = compute_sentence_impact(shifters);
           }
         }
       }
-
-      update_token_scores_sentences(scores, tokenScores, normalizer, nPolarized, tokenShifters, tokenWeights, nL, nTokens, how, nPuncts);
+      update_token_scores(scores, tokenScores, normalizer, nPolarized, tokenShifters, tokenWeights, nL, nTokens, how, nPuncts);
 
       sentScores(i, 0) = nTokens - nPuncts;
       for (int m = 0; m < nL; m++) {